requirement.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
djangorestframework==3.14.0
dnspython==2.4.1
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
multidict==6.0.4
packaging==23.1
pycryptodome==3.10.1
pymongo==4.4.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2